refactor(docs): ownership-based naming on /scores (#58)
