+++ apps/web/src/routes/scores/scores-leaderboard-graph-controls.svelte
