+++ apps/web/src/routes/scores/scores-leaderboard-layout-constants.ts
