+++ apps/web/src/routes/scores/scores-leaderboard-scroll-fades.svelte
