+++ apps/web/src/routes/scores/scores-leaderboard-scroll-state.svelte.ts
@@ -3,20 +3,20 @@ import { onMount } from 'svelte'
 import {
   getEmptyGraphVisibility,
   getGraphVisibility,
-} from './scores-data-helpers'
+} from './scores-leaderboard-data-transforms'
 import {
   SCORE_DIAGONAL_OVERFLOW_PX,
   SCORE_ROW_HEIGHT_FULL_PX,
   SCORE_SCROLL_RESET_DELAY_MS,
   SCORE_VIRTUAL_OVERSCAN,
-} from './scores-layout-constants'
+} from './scores-leaderboard-layout-constants'
 import type {
   CurrentUserScoreData,
   GraphVisibility,
   ScoreEntry,
   ScoreGraphEntry,
   ViewportVisibility,
-} from './types'
+} from './scores-shared-types'
 
 type ScoresScroll = ReturnType<typeof useInfiniteVirtualScroll>
 type SelfRowPosition = 'top' | 'bottom'
