+++ apps/web/src/routes/scores/scores-leaderboard-team-row-delta-indicator.svelte
