+++ apps/web/src/routes/scores/scores-leaderboard-team-row-sparkline.svelte
