+++ apps/web/src/routes/scores/scores-leaderboard.svelte
@@ -2,14 +2,14 @@
   import { ScrollArea } from '$lib/components'
   import type { Snippet } from 'svelte'
   import type { Attachment } from 'svelte/attachments'
-  import ScoresChallengeHeader from './scores-challenge-header.svelte'
-  import type { createScoresDataModel } from './scores-data-model.svelte'
-  import ScoresFades from './scores-fades.svelte'
-  import ScoresGraphControls from './scores-graph-controls.svelte'
-  import ScoresGraph from './scores-graph.svelte'
-  import type { createScoresRouteState } from './scores-route-state.svelte'
-  import type { createScoresViewportState } from './scores-viewport-state.svelte'
-  import type { ChallengeInfo, ScoreGraphEntry } from './types'
+  import ScoresChallengeHeader from './scores-leaderboard-challenge-header.svelte'
+  import type { createScoresDataModel } from './scores-leaderboard-data-model.svelte'
+  import ScoresGraphControls from './scores-leaderboard-graph-controls.svelte'
+  import ScoresGraph from './scores-leaderboard-graph.svelte'
+  import ScoresFades from './scores-leaderboard-scroll-fades.svelte'
+  import type { createScoresViewportState } from './scores-leaderboard-scroll-state.svelte'
+  import type { createScoresRouteState } from './scores-page-url-state.svelte'
+  import type { ChallengeInfo, ScoreGraphEntry } from './scores-shared-types'
 
   type ScoreData = ReturnType<typeof createScoresDataModel>
   type RouteState = ReturnType<typeof createScoresRouteState>
