+++ apps/web/src/routes/scores/scores-page-toolbar-search-box.svelte
