+++ apps/web/src/routes/scores/scores-page-url-params.ts
@@ -1,4 +1,4 @@
-import type { SortMode, ViewMode } from './types'
+import type { SortMode, ViewMode } from './scores-shared-types'
 
 export const MIN_SEARCH_LENGTH = 2
