+++ apps/web/src/routes/scores/scores-screenshot-export-modal.svelte
@@ -15,8 +15,8 @@
   import { toast } from 'svelte-sonner'
   import ScoresScreenshotPreview, {
     type ScreenshotOptions,
-  } from './scores-screenshot-preview.svelte'
-  import type { CategoryGroup } from './types'
+  } from './scores-screenshot-export-preview.svelte'
+  import type { CategoryGroup } from './scores-shared-types'
 
   interface TeamEntry {
     id: string
