+++ apps/web/src/routes/scores/scores-shared-types.ts
